prompt before installing anything

+++ init/10_ubuntu.sh
@@ -1,6 +1,9 @@
 # Ubuntu-only stuff. Abort if not Ubuntu.
 [[ "$(cat /etc/issue 2> /dev/null)" =~ Ubuntu ]] || return 1
 
+#only run if local
+[[ ! "$SSH_TTY" ]] || return 1
+
 # If the old files isn't removed, the duplicate APT alias will break sudo!
 sudoers_old="/etc/sudoers.d/sudoers-cowboy"; [[ -e "$sudoers_old" ]] && sudo rm "$sudoers_old"
 
@@ -30,39 +33,49 @@ EOF
   fi
 fi
 
-# Update APT.
-e_header "Updating APT"
-sudo apt-get -qq update
-sudo apt-get -qq dist-upgrade
 
-# Install APT packages.
-packages=(
-  build-essential libssl-dev
-  git-core
-  tree sl id3tool cowsay
-  nmap telnet
-  htop
-)
+cat <<EOF
+This installation of Ubuntu can be updated and a bunch of packages can be installed.  Should we?
+This will be skipped if "Y" isn't pressed within the next 15 seconds.
+EOF
+read -N 1 -t 15 -p "Update Ubuntu? [y/N] " update_ubuntu; echo
+if [[ "$update_ubuntu" =~ [Yy] ]]; then
+  e_header "Updating Ubuntu"
 
-list=()
-for package in "${packages[@]}"; do
-  if [[ ! "$(dpkg -l "$package" 2>/dev/null | grep "^ii  $package")" ]]; then
-    list=("${list[@]}" "$package")
-  fi
-done
+  # Update APT.
+  e_header "Updating APT"
+  sudo apt-get -qq update
+  sudo apt-get -qq dist-upgrade
 
-if (( ${#list[@]} > 0 )); then
-  e_header "Installing APT packages: ${list[*]}"
-  for package in "${list[@]}"; do
-    sudo apt-get -qq install "$package"
-  done
-fi
-
-# Install Git Extras
-if [[ ! "$(type -P git-extras)" ]]; then
-  e_header "Installing Git Extras"
-  (
-    cd ~/.dotfiles/libs/git-extras &&
-    sudo make install
+  # Install APT packages.
+  packages=(
+    build-essential libssl-dev
+    git-core
+    tree sl id3tool cowsay
+    nmap telnet
+    htop
   )
+
+  list=()
+  for package in "${packages[@]}"; do
+    if [[ ! "$(dpkg -l "$package" 2>/dev/null | grep "^ii  $package")" ]]; then
+      list=("${list[@]}" "$package")
+    fi
+  done
+
+  if (( ${#list[@]} > 0 )); then
+    e_header "Installing APT packages: ${list[*]}"
+    for package in "${list[@]}"; do
+      sudo apt-get -qq install "$package"
+    done
+  fi
+
+  # Install Git Extras
+  if [[ ! "$(type -P git-extras)" ]]; then
+    e_header "Installing Git Extras"
+    (
+      cd ~/.dotfiles/libs/git-extras &&
+      sudo make install
+    )
+  fi
 fi
